vm: Make vm_fault_(quick_)hold_pages take a pointer
Rather than passing a virtual address (usually cast from a pointer),
pass a pointer to vm_fault_quick_hold_pages as the starting address.

Alter most callers to also take void * arguments and bump
__FreeBSD_version to allow external code (e.g., OpenZFS) to be updated.
The one exception in linuxkpi which I've left alone except for a cast at
the call site for now.  CHERI Linux currently leaves these APIs passing
unsigned long so we'll take a compatability hit when that day comes.

With CHERI, this allows the bounds and permission to be checked before
mapping the memory in question for use by the caller.  Using (void *)
is prefered to using uintptr_t because it makes it clear that this is
always a pointer rather than allowing confusion with vm_offset_t.

Effort:		CHERI upstreaming
Sponsored by:	Innovate UK