share/mk: basic support for group options
Support group options where 1 of n values will be selected (or a default
value will be used). After processing an OPT_FOO will be set to one
value from FOO_OPTIONS for each FOO in __SINGLE_OPTIONS.  If the user
sets FOO that value will be used, otherwise __FOO_DEFAULT will be used
(if __FOO_DEFAULT is not explicitly defined it will be set to the first
value in __FOO_OPTIONS).

For options that don't work on a particular system,
BROKEN_SINGLE_OPTIONS can be set to a list of 3-tuples of the form:
	OPTION broken_value replacement_value

This is somewhat inspired by OPTIONS_SINGLE from ports, but the
structure is quite different with a per-option variable a la MK_FOO.