libc: stop exposing break/sbrk/brk
These interfaces are convoluted, rely on memory layout details that
should not be exposed, and are often misued as proxies for other things
(e.g., calling sbrk(0) repeatedly to determing heap use despite modern
allocators prefering mmap.)  They are not part of any standard.  While
we must continue to support them for backwards compatability, most uses
can be removed in a straightfoward manner.

Declare the undocumented break(2) system call COMPAT14 and remove public
decleration and default version of it, sbrk, and brk interfaces.  A
later commit will introduce a libsbrk to allow their use for programs
that can not be altered to avoid them (due to the way that autoconf
works, the symbol must not be directly linkable to avoid it being found
and used even without a decleration).