libc: use __sys___realpathat directly in realpath
We don't need to use an interposable symbol for this purpose and it's
simpler to just call the syscall in libsys.  The to another bug (fixed
in the next commit) we were incorrectly using __realpathat in libc not
libsys.

While here, drop support for running on a FreeBSD 12 kernel and simplify
includes.