libsys: Don't create or expose __realpathat
Previously, __realpathat was in libsys, but not exported from libc which
meant the stub in libc was not filtered and thus libc's copy of the
syscall was used.  This broke an upcoming change to CheriBSD limiting
syscalls to libsys.

The realpath(3) implementation now uses __sys___realpathat so there are no
consumers of __realpathat.  Switch it to PSEUDO (only _foo and __sys_foo
symbols) and remove __realpathat from Symbol.map.