Add /etc/rc.d/newaliases
This script runs on firstboot and runs newaliases if /etc/mail/aliases
exists and /etc/mail/aliases.db does not.  This allows image creation
scripts to avoid running newaliases.

Due to the switch to dma as the default mailier, this is a no-op in
practice unless the image has configured an alternate mailer.

/etc/rc.d/sendmail has the ability to create/update aliases.db, but it
is disabled by default even if sendmail is enabled.