csu: drop support for GCJ
Remove .jcr sections and related support.  This has no impact on
existing binaries but prevents the creation of new ones.

GCC removed support from trunk in on September 30, 2016 and the last
release with support with GCC 6.5 so we don't even have a port
capability of producing such binaries.

For reference see:
    https://en.wikipedia.org/wiki/GNU_Compiler_for_Java#History