reallocf: fix a leak if realloc exhibits SysV behavior
If realloc() is (mis)implemented to follow SysV's behavior of
realloc(ptr, 0) being equivalent to free(ptr) we need to check errno
before calling free() to distinguish between the pointer being freed and
failure in order to avoid a double free().

With our current jemalloc implementation this is a non-issue and the
comment refered to pkh-malloc and an early jemalloc, but malloc() can
be replaced via LD_PRELOAD and a single load and conditional branch in
reallocf() will never be performance limiting.

For futher discussion see:
   https://www.openwall.com/lists/musl/2025/06/21/19

Reported by:	Alejandro Colomar <alx@kernel.org>