libc: remove use of the _<syscall> abstraction
Back in 2001 (commit d201fe46e3552), libc was made thread safe and
the folling convention for syscalls was introduced:
	__sys_foo - actual system call
	_foo - weak definition to __sys_foo
	foo - weak definition to __sys_foo

libc was then changed to use _foo for all system calls wrapped by
the userspace threading library and the namespace.h and un-namespace.h
headers were used to produce declerations from standard headers.  This
was done to support libc_r style userspace threading and no longer
serves a purpose so remove the use of this abstraction in favor of using
the (mostly) generated libsys.h and __sys_foo symbols.

Because syscalls (other than syscall(2) and __syscall(2)) aren't
actually variadic this has a couple side effects:
 - open(at) always take a mode argument
 - ioctl always takes a char * argument so we cast all arguments

I made this change with a combination of scripting to remove most
includes of (un-)namespace.h, a Coccinelle script (added to
tools/coccinelle as an example), and some manual editing to add libsys.h
includes and deal with cases where Coccinelle's patching didn't match
style(9) or local style.