sendfile(2): rework compat implementation
Rename the sendfile() function to kern_sendfile(), expand the arguments
in sendfile_args, and extend with two function pointer arguments to
copy in the header/trailer structure and the create uio's for the
header and trailer as required. Use this to allow the removal of
freebsd32_do_sendfile() which was a nearly identical duplicate of
sendfile() with attendant maintenance cost.

Effort:		CHERI upstreaming
Sponsored by:	Innovate UK